Mudando para BigDecimal. Criando Model Tabela.

+++ projeto-alimentos/src/main/java/beanDTO/GraficoMacros.java
@@ -1,47 +1,46 @@
 package beanDTO;
 
+import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.Date;
 import java.util.List;
 
 public class GraficoMacros {
 	List <Date> listaData=new ArrayList<Date>();
-	List <Double> listaCalorias=new ArrayList<Double>();
-	List <Double> listaProteinas=new ArrayList<Double>();
-	List <Double> listaCarboidratos=new ArrayList<Double>();
-	List <Double> listaGorduras=new ArrayList<Double>();
+	List <BigDecimal> listaCalorias=new ArrayList<BigDecimal>();
+	List <BigDecimal> listaProteinas=new ArrayList<BigDecimal>();
+	List <BigDecimal> listaCarboidratos=new ArrayList<BigDecimal>();
+	List <BigDecimal> listaGorduras=new ArrayList<BigDecimal>();
 	public List<Date> getListaData() {
 		return listaData;
 	}
 	public void setListaData(List<Date> listaData) {
 		this.listaData = listaData;
 	}
-	public List<Double> getListaCalorias() {
+	public List<BigDecimal> getListaCalorias() {
 		return listaCalorias;
 	}
-	public void setListaCalorias(List<Double> listaCalorias) {
+	public void setListaCalorias(List<BigDecimal> listaCalorias) {
 		this.listaCalorias = listaCalorias;
 	}
-	public List<Double> getListaProteinas() {
+	public List<BigDecimal> getListaProteinas() {
 		return listaProteinas;
 	}
-	public void setListaProteinas(List<Double> listaProteinas) {
+	public void setListaProteinas(List<BigDecimal> listaProteinas) {
 		this.listaProteinas = listaProteinas;
 	}
-	public List<Double> getListaCarboidratos() {
+	public List<BigDecimal> getListaCarboidratos() {
 		return listaCarboidratos;
 	}
-	public void setListaCarboidratos(List<Double> listaCarboidratos) {
+	public void setListaCarboidratos(List<BigDecimal> listaCarboidratos) {
 		this.listaCarboidratos = listaCarboidratos;
 	}
-	public List<Double> getListaGorduras() {
+	public List<BigDecimal> getListaGorduras() {
 		return listaGorduras;
 	}
-	public void setListaGorduras(List<Double> listaGorduras) {
+	public void setListaGorduras(List<BigDecimal> listaGorduras) {
 		this.listaGorduras = listaGorduras;
 	}
-	
-	
-	
+
 
 }
